+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Loading state when the query is being made.
+
+### Changed
+- Make the query of supported languages only when the user interacts with the `LocaleSwitcher`.
 
 ## [0.5.4] - 2020-02-13
 ### Added